+++ ui/src/util/BrowserUtil.ts
@@ -4,13 +4,20 @@ export default class BrowserUtil {
 
   static tryLocalStorageSetItem(key: string, value: string): boolean {
     if (window === undefined || window.localStorage === undefined) return false;
-    window.localStorage.setItem(key, value);
+    try {
+      window.localStorage.setItem(key, value);
+    } catch {
+      return false;
+    }
     return true;
   }
 
   static tryLocalStorageGetItem(key: string, defaultValue: any): any {
     if (window === undefined || window.localStorage === undefined)
       return defaultValue;
-    return window.localStorage.getItem(key) ?? defaultValue;
+    try {
+      return window.localStorage.getItem(key) ?? defaultValue;
+    } catch {}
+    return defaultValue;
   }
 }